refactor(runtime): remove ILifecycleTask
